fix(error-tracking): simplify issue filter previews
