+++ products/error_tracking/frontend/components/Breakdowns/consts.ts
@@ -1,15 +1,12 @@
 export interface BreakdownPreset {
     property: string
     title: string
-    removable?: boolean
 }
 
 export const LIMIT_ITEMS = 3
 export const BREAKDOWN_DETAILS_LIMIT = 100
 
-// Each breakdown property adds an arrayJoin fan-out to the query, so the auto-derived list from the
-// selected event is capped to keep the page-load cost from scaling with the event's schema width.
-// Any property beyond the cap stays reachable through the property picker.
+// Each property adds an arrayJoin fan-out, so cap event-derived properties to keep query cost bounded.
 export const MAX_SELECTED_EVENT_BREAKDOWN_PROPERTIES = 10
 
 export const BREAKDOWN_PRESETS: BreakdownPreset[] = [
@@ -23,4 +20,5 @@ export const BREAKDOWN_PRESETS: BreakdownPreset[] = [
 
 export const BreakdownsEvents = {
     MiniBreakdownsLoaded: 'error tracking mini breakdowns loaded',
+    MiniBreakdownsPropertySelected: 'error tracking mini breakdowns property selected',
 }
