+++ products/error_tracking/frontend/components/Breakdowns/miniBreakdownsLogic.ts
@@ -88,23 +88,11 @@ export function getBreakdownForPropertyFromResponse(
     return { properties, totalCount }
 }
 
-export function buildBreakdownProperties(
-    customBreakdownProperties: string[],
-    selectedEventBreakdownProperties: BreakdownPreset[]
-): BreakdownPreset[] {
-    const configuredProperties = [
-        ...BREAKDOWN_PRESETS,
-        ...customBreakdownProperties.map((property) => ({
-            property,
-            title: getBreakdownPropertyTitle(property),
-            removable: true,
-        })),
-    ]
-    const configuredPropertyNames = new Set(configuredProperties.map(({ property }) => property))
-
+export function buildBreakdownProperties(selectedEventBreakdownProperties: BreakdownPreset[]): BreakdownPreset[] {
+    const presetPropertyNames = new Set(BREAKDOWN_PRESETS.map(({ property }) => property))
     return [
-        ...configuredProperties,
-        ...selectedEventBreakdownProperties.filter(({ property }) => !configuredPropertyNames.has(property)),
+        ...BREAKDOWN_PRESETS,
+        ...selectedEventBreakdownProperties.filter(({ property }) => !presetPropertyNames.has(property)),
     ]
 }
 
@@ -118,12 +106,10 @@ export interface miniBreakdownsLogicValues {
     breakdownDetailsError: string | null
     breakdownDetailsLoading: boolean
     breakdownProperties: BreakdownPreset[]
-    customBreakdownProperties: string[]
     getBreakdownForProperty: (property: string) => {
         properties: BreakdownSinglePropertyStat[]
         totalCount: number
     }
-    propertyPickerOpenRequest: number
     response: ErrorTrackingBreakdownsQueryResponse | null
     responseError: string | null
     responseLoading: boolean
@@ -133,9 +119,6 @@ export interface miniBreakdownsLogicValues {
 
 // Generated by kea-typegen. Update if you're an agent, ignore if you're human.
 export interface miniBreakdownsLogicActions {
-    addBreakdownProperty: (property: string) => {
-        property: string
-    }
     closeBreakdownDetails: () => {
         value: true
     }
@@ -160,7 +143,9 @@ export interface miniBreakdownsLogicActions {
             property: string
         }
     }
-    loadResponse: () => any
+    loadResponse: () => {
+        value: true
+    }
     loadResponseFailure: (
         error: string,
         errorObject?: any
@@ -170,19 +155,17 @@ export interface miniBreakdownsLogicActions {
     }
     loadResponseSuccess: (
         response: ErrorTrackingBreakdownsQueryResponse,
-        payload?: any
+        payload?: {
+            value: true
+        }
     ) => {
         response: ErrorTrackingBreakdownsQueryResponse
-        payload?: any
+        payload?: {
+            value: true
+        }
     }
     openBreakdownDetails: (breakdown: BreakdownPreset) => {
         breakdown: BreakdownPreset
-    }
-    openPropertyPicker: () => {
-        value: true
-    }
-    removeBreakdownProperty: (property: string) => {
-        property: string
     }
 }
 
@@ -194,10 +177,7 @@ export interface miniBreakdownsLogicMeta {
             selectedEvent: ErrorEventType | null,
             isCloudOrDev: boolean | undefined
         ) => BreakdownPreset[]
-        breakdownProperties: (
-            customBreakdownProperties: string[],
-            selectedEventBreakdownProperties: BreakdownPreset[]
-        ) => BreakdownPreset[]
+        breakdownProperties: (selectedEventBreakdownProperties: BreakdownPreset[]) => BreakdownPreset[]
         getBreakdownForProperty: (response: ErrorTrackingBreakdownsQueryResponse | null) => (property: string) => {
             properties: BreakdownSinglePropertyStat[]
             totalCount: number
@@ -227,30 +207,11 @@ export const miniBreakdownsLogic = kea<miniBreakdownsLogicType>([
         ],
     })),
     actions({
-        addBreakdownProperty: (property: string) => ({ property }),
         closeBreakdownDetails: true,
+        loadResponse: true,
         openBreakdownDetails: (breakdown: BreakdownPreset) => ({ breakdown }),
-        removeBreakdownProperty: (property: string) => ({ property }),
-        openPropertyPicker: true,
     }),
     reducers({
-        customBreakdownProperties: [
-            [] as string[],
-            {
-                addBreakdownProperty: (properties, { property }) =>
-                    BREAKDOWN_PRESETS.some((preset) => preset.property === property) || properties.includes(property)
-                        ? properties
-                        : [...properties, property],
-                removeBreakdownProperty: (properties, { property }) =>
-                    properties.filter((candidate) => candidate !== property),
-            },
-        ],
-        propertyPickerOpenRequest: [
-            0,
-            {
-                openPropertyPicker: (request) => request + 1,
-            },
-        ],
         selectedBreakdownProperty: [
             null as BreakdownPreset | null,
             {
@@ -280,8 +241,8 @@ export const miniBreakdownsLogic = kea<miniBreakdownsLogicType>([
         breakdownDetails: [
             null as ErrorTrackingBreakdownsQueryResponse | null,
             {
-                loadBreakdownDetails: async ({ property }: { property: string }) =>
-                    await api.query(
+                loadBreakdownDetails: async ({ property }: { property: string }, breakpoint) => {
+                    const result = await api.query(
                         errorTrackingBreakdownsQuery({
                             issueId: props.issueId,
                             dateRange: values.dateRange,
@@ -290,13 +251,17 @@ export const miniBreakdownsLogic = kea<miniBreakdownsLogicType>([
                             maxValuesPerProperty: BREAKDOWN_DETAILS_LIMIT,
                         }),
                         { refresh: 'blocking' }
-                    ),
+                    )
+                    breakpoint()
+                    return result
+                },
             },
         ],
         response: [
             null as ErrorTrackingBreakdownsQueryResponse | null,
             {
-                loadResponse: async () => {
+                loadResponse: async (_: { value: true }, breakpoint) => {
+                    await breakpoint(100)
                     const startTime = Date.now()
                     // Read connected values before the request because either connected logic may unmount while it runs.
                     const { dateRange, filterTestAccounts } = values
@@ -311,6 +276,7 @@ export const miniBreakdownsLogic = kea<miniBreakdownsLogicType>([
                         }),
                         { refresh: 'blocking' }
                     )
+                    breakpoint()
 
                     posthog.capture(BreakdownsEvents.MiniBreakdownsLoaded, {
                         issueId: props.issueId,
@@ -333,9 +299,9 @@ export const miniBreakdownsLogic = kea<miniBreakdownsLogicType>([
                 getSelectedEventBreakdownProperties(selectedEvent?.properties ?? null, isCloudOrDev),
         ],
         breakdownProperties: [
-            (selectors) => [selectors.customBreakdownProperties, selectors.selectedEventBreakdownProperties],
-            (customBreakdownProperties: string[], selectedEventBreakdownProperties: BreakdownPreset[]) =>
-                buildBreakdownProperties(customBreakdownProperties, selectedEventBreakdownProperties),
+            (selectors) => [selectors.selectedEventBreakdownProperties],
+            (selectedEventBreakdownProperties: BreakdownPreset[]) =>
+                buildBreakdownProperties(selectedEventBreakdownProperties),
         ],
         getBreakdownForProperty: [
             (selectors) => [selectors.response],
@@ -364,9 +330,7 @@ export const miniBreakdownsLogic = kea<miniBreakdownsLogicType>([
         },
     })),
     listeners(({ actions }) => ({
-        addBreakdownProperty: () => actions.loadResponse(),
         openBreakdownDetails: ({ breakdown }) => actions.loadBreakdownDetails({ property: breakdown.property }),
-        removeBreakdownProperty: () => actions.loadResponse(),
     })),
     events(({ actions }) => ({
         afterMount: () => {
